make mxnet optional
