+++ text/src/autogluon/text/__init__.py
@@ -1,6 +1,6 @@
 from . import automm
 from . import text_prediction
 from .text_prediction import TextPredictor
-from .text_prediction.presets import ag_text_presets, list_presets
+from .text_prediction.presets import list_text_presets
 
 from .version import __version__
