+++ text/src/autogluon/text/text_prediction/__init__.py
@@ -3,11 +3,10 @@ from . import infer_types
 from . import metrics
 from . import predictor
 from . import presets
-from . import text_presets
 from .predictor import TextPredictor
 
 __all__ = [
     'constants', 'infer_types', 'metrics',
-    'predictor', 'presets', 'text_presets',
+    'predictor', 'presets',
     'TextPredictor',
 ]
