+++ text/src/autogluon/text/text_prediction/mx_predictor.py
@@ -15,7 +15,7 @@ from autogluon.core.utils.miscs import in_ipynb
 
 
 from . import constants as _C
-from .presets import ag_text_presets, merge_params
+from .legacy_presets import ag_text_presets, merge_params
 from .infer_types import infer_column_problem_types, printable_column_type_string
 from .metrics import infer_eval_log_metrics
 from .. import version
